Use the correct CSS selector to find out if names are required
The selector `#register_names` doesn't seem to be in the DOM.

This change replaces that check with one that checks for the
value of the hidden input named `action`. If that value is
`register_names` we make sure to collect the names and add
them to the AJAX response to the server.

Fixes #86